Remove required fields from childs to create reservation's DTO
